moved finding local media files process to repository class
