+++ app/src/main/java/com/freshdigitable/ohlplayer/MediaListActivity.kt
@@ -1,21 +1,14 @@
 package com.freshdigitable.ohlplayer
 
-import android.content.ContentUris
 import android.content.pm.PackageManager
-import android.media.MediaMetadataRetriever
-import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.os.Environment
-import android.provider.MediaStore
-import android.util.Log
 import android.util.Size
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.DiffUtil
@@ -26,10 +19,8 @@ import com.freshdigitable.ohlplayer.databinding.ActivityMediaListBinding
 import com.freshdigitable.ohlplayer.databinding.ViewMediaListItemBinding
 import com.freshdigitable.ohlplayer.model.PlayableItem
 import com.freshdigitable.ohlplayer.store.PlayableItemStore
-import com.freshdigitable.ohlplayer.store.create
 import com.freshdigitable.ohlplayer.store.uri
 import kotlinx.coroutines.flow.collectLatest
-import java.io.File
 import java.io.IOException
 
 class MediaListActivity : AppCompatActivity() {
@@ -45,14 +36,9 @@ class MediaListActivity : AppCompatActivity() {
         val linearLayoutManager = LinearLayoutManager(applicationContext)
         linearLayoutManager.isAutoMeasureEnabled = true
         binding.list.layoutManager = linearLayoutManager
-    }
 
-    override fun onStart() {
-        super.onStart()
-        playableItemStore.open()
         val adapter = ViewAdapter(playableItemStore)
-        addLocalMediaFiles()
-        binding?.list?.adapter = adapter
+        binding.list.adapter = adapter
         lifecycleScope.launchWhenCreated {
             playableItemStore.getMediaItems().collectLatest {
                 adapter.submitList(it)
@@ -68,107 +54,14 @@ class MediaListActivity : AppCompatActivity() {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         if (requestCode == 100) {
             if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                addLocalMediaFiles()
+                playableItemStore.findLocalMediaFiles()
             }
         }
     }
 
-    private fun addLocalMediaFiles() {
-        val playableItems = listOf(Environment.DIRECTORY_MUSIC, Environment.DIRECTORY_MOVIES)
-            .flatMap { findPlayableItems(it) }
-        playableItemStore.registerIfAbsent(playableItems)
-    }
-
-    private fun findPlayableItems(type: String): List<PlayableItem> {
-        val files = findMediaFiles(type)
-        if (files.isEmpty()) {
-            return emptyList()
-        }
-        return MediaMetadataRetriever().use { metadataRetriever ->
-            files.map { uri ->
-                metadataRetriever.setDataSource(applicationContext, uri)
-                val title =
-                    metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_TITLE)
-                val artist =
-                    metadataRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ARTIST)
-                val path = uri.toString()
-                PlayableItem.create(
-                    path = path,
-                    title = if (title.isNullOrEmpty()) File(path).name else title,
-                    artist = artist,
-                )
-            }
-        }
-    }
-
-    private fun findMediaFiles(type: String): List<Uri> {
-        Log.d("MediaListActivity", "findPlayableFiles: $type")
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            getFileListQ(type)
-        } else {
-            getFileList(type)
-        }
-    }
-
-    private fun getFileList(type: String): List<Uri> {
-        val externalFilesDir = Environment.getExternalStoragePublicDirectory(type)
-        val fileList = externalFilesDir.list() ?: return emptyList()
-        return fileList.map { File(externalFilesDir, it) }
-            .filter { it.isFile }
-            .filter { playableItemStore.findByPath(it.absolutePath) == null }
-            .map { Uri.fromFile(it) }
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.Q)
-    private fun getFileListQ(type: String): List<Uri> {
-        val collection: Uri = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            when (type) {
-                Environment.DIRECTORY_MUSIC ->
-                    MediaStore.Audio.Media.getContentUri(MediaStore.VOLUME_EXTERNAL)
-                Environment.DIRECTORY_MOVIES ->
-                    MediaStore.Video.Media.getContentUri(MediaStore.VOLUME_EXTERNAL)
-                else -> throw IllegalArgumentException()
-            }
-        } else {
-            when (type) {
-                Environment.DIRECTORY_MUSIC -> MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
-                Environment.DIRECTORY_MOVIES -> MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-                else -> throw IllegalArgumentException()
-            }
-        }
-        val projection = arrayOf(
-            MediaStore.Video.Media._ID,
-            MediaStore.MediaColumns.RELATIVE_PATH
-        )
-        applicationContext.contentResolver.query(
-            collection,
-            projection,
-            null,
-            null,
-            null
-        ).use { cursor ->
-            // Cache column indices.
-            val idColumn = cursor!!.getColumnIndexOrThrow(MediaStore.Video.Media._ID)
-            val relativePathColumn =
-                cursor.getColumnIndexOrThrow(MediaStore.MediaColumns.RELATIVE_PATH)
-            val files = mutableListOf<Uri>()
-            while (cursor.moveToNext()) {
-                val id = cursor.getLong(idColumn)
-                val relativePath = cursor.getString(relativePathColumn)
-                if (relativePath.startsWith("Music/") || relativePath.startsWith("Video/")) {
-                    val contentUri = ContentUris.withAppendedId(collection, id)
-                    Log.d("MediaListActivity", "getFileListQ: $contentUri, $relativePath")
-                    files.add(contentUri)
-                }
-            }
-            return files
-        }
-    }
-
-    override fun onStop() {
-        super.onStop()
+    override fun onDestroy() {
+        super.onDestroy()
         binding?.list?.adapter = null
-        playableItemStore.close()
     }
 
     companion object {
